Fixed up the gulp pipeline so that it include $ jekyll serve. Now to autoprefix and then serve the website in one command, use $ gulp.
Fixed navigation bar image sizing issues. Am semi-happy with the navigation bar design now. Would like to put to logo on the left when screen min-width is >800px in the future.

Decided on navigation bar items and made landing pages for each.

Attempting to put pages into a flexbox on home page.